requirements.txt
django-webtest==1.9.2
dj-database-url==0.4.1
Django==2.0rc1
django-static-precompiler==1.5
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gunicorn==19.6.0
lxml==4.1.1
olefile==0.43
pbr==3.1.1
django-fsm==2.6.0
psycopg2==2.6.2
pep8==1.7.0
Pillow==4.0.0
PyJWT==1.5.3
pytz==2017.3
reportlab==3.3.0
six==1.11.0
stevedore==1.28.0
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
whitenoise==3.3.1
xlrd==1.1.0
